Refactor: Rename data package structure to improve organization and clarity
