+++ app/src/main/java/com/example/sheetcompute/domain/database/paging/AttendanceSummaryPagingSource.kt
@@ -1,12 +1,10 @@
-package com.example.sheetcompute.data.paging
+package com.example.sheetcompute.domain.database.paging
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.example.sheetcompute.data.daos.AttendanceDao
+import com.example.sheetcompute.domain.database.daos.AttendanceDao
 import com.example.sheetcompute.data.entities.AttendanceRecordUI
 import com.example.sheetcompute.data.entities.AttendanceSummary
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.first
 import java.time.LocalDate
 
 class AttendanceSummaryPagingSource(
