+++ app/src/main/java/com/example/sheetcompute/domain/database/roomDB/AppDatabase.kt
@@ -1,12 +1,12 @@
-package com.example.sheetcompute.data.roomDB
+package com.example.sheetcompute.domain.database.roomDB
 
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.example.sheetcompute.data.daos.AttendanceDao
-import com.example.sheetcompute.data.daos.EmployeeDao
-import com.example.sheetcompute.data.daos.HolidayDao
+import com.example.sheetcompute.domain.database.daos.AttendanceDao
+import com.example.sheetcompute.domain.database.daos.EmployeeDao
+import com.example.sheetcompute.domain.database.daos.HolidayDao
 import com.example.sheetcompute.data.entities.AttendanceRecord
 import com.example.sheetcompute.data.entities.EmployeeEntity
 import com.example.sheetcompute.data.entities.Holiday
