+++ app/src/main/java/com/example/sheetcompute/domain/repo/EmployeeRepo.kt
@@ -1,7 +1,7 @@
 package com.example.sheetcompute.domain.repo
 
 import com.example.sheetcompute.data.entities.EmployeeEntity
-import com.example.sheetcompute.data.roomDB.AppDatabase
+import com.example.sheetcompute.domain.database.roomDB.AppDatabase
 
 class EmployeeRepo {
     private val database by lazy { AppDatabase.get() }
